Intervals minimal length check
While using lt i noticed that due to nearest-integer-approximations that the algorythm uses some intervals were indeed 0 days intervals thus compromising future calculations (new intervals are calculated by multiplying the current interval and if that is zero...) and also adding flashcards to review today without notifying the user.

Now if a 0-days intervals is found it is immediately transformed in a 1-day interval.