Autobackup yout deck
Since I'm not really confident about lt and given the way it acts on the deckfiles I added an autobackup feature that automatically backup your deck in a backup_ file in the same directory just in case some weir things happens during the execution of the reviews. I am also using it quite a bit recently and as my decks grow older I feel the need to be sure that all my data would not be entirely destroyed in the blink of an eye; lt automatic backups are an easy way to prevent this possibility.