More info
Added some more info to the info command. Now prints also the average E-factor, considering mapping the E-factor average to the [0,1] interval since it never goes below 1.3 nor above 2.5 and it can be a bit confusing for the user that does not know how SM2 works (however I always assume my userbase to be as autistic as me so this "fix" will probably never become a thing).