Version v0.1.0
Fixed a sneaky bug that would have potentially caused problems with the coming of a new year (most likely 2019) and that for some reasons stuck me in a loop in which I ended up with several flashcards scheduled for 2026. IN order to fix it I had to change the date format I was using to store flashcards due dates and so I thought to release a first version not much because I think nclt has reached a pretty stable point (which I do, btw) but mostly as a separation from the old date format.